requirements.txt
aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
exceptiongroup==1.2.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.6
loguru==0.7.3
magic-filter==1.0.12
MarkupSafe==2.1.4
multidict==6.0.4
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
python-multipart==0.0.19
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
yarl==1.9.4
